Question one
We will need a table to keep track of user clicks. Below is a rough structure of the table, but it is incomplete in the following ways:

There are no primary keys,
More columns may need to be added,
The data types are not all specified.
Copy the table below, paste it into your response, and complete the missing parts of the schema according to the list above. You may provide a brief explanation as needed.

Explanation: 
UserID which is underlined because it is the primary key as the principal means of identifying tuples within a relation. 
Each data types are specified.
Since users are within a team, team ID and team level are included in the clicks data.
Grid cell and flamingo subtype specify another attribute of data related to user click.